Add support for character constants

+++ examples/hello.rs
@@ -4,7 +4,7 @@ use rhai::Engine;
 fn main() {
     let mut engine = Engine::new();
 
-    if let Ok(result) = engine.eval::<i32>("40 + 2") {
+    if let Ok(result) = engine.eval::<i64>("40 + 2") {
         println!("Answer: {}", result);  // prints 42
     }
 }
